requirements.txt
av2==0.3.5
diffusers==0.33.1
matplotlib==3.6.1
numba==0.58.1
numpy==1.23.0
pandas==1.5.3
protobuf==3.20.3
pytorch-lightning==2.4.0
shapely==2.1.1
tensorboard==2.12.3
tensorflow==2.12.0
torch==2.1.0
torch-geometric==2.6.1
tqdm
waymo-open-dataset-tf-2-12-0==1.6.4
